requirements.txt
Django==5.1.4
Pillow==10.1.0
qrcode==7.4.2
xlsxwriter==3.1.9 